Referal Changes

+++ src/main/java/com/maan/eway/bean/CompanyProductMaster.java
@@ -19,6 +19,7 @@ import org.hibernate.annotations.DynamicUpdate;
 import lombok.*;
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 import java.util.Date;
@@ -123,10 +124,10 @@ private static final long serialVersionUID = 1L;
     private String     regulatoryCode ;
 
     @Column(name="SUM_INSURED_START")
-    private Double     sumInsuredStart ;
+    private BigDecimal     sumInsuredStart ;
 
     @Column(name="SUM_INSURED_END")
-    private Double     sumInsuredEnd ;
+    private BigDecimal     sumInsuredEnd ;
 
     @Column(name="COMMISSION_VAT_YN", length=1)
     private String     commissionVatYn ;
